Errors in F RSSAC stats
To: "Wessels, Duane" <dwessels@verisign.com>
From: Ray Bellis <ray@isc.org>
Subject: Errors in F RSSAC stats
Date: Tue, 31 Aug 2021 13:44:49 +0100

We had a data processing issue in August, affecting data from the 8th to
24th.

The slightly longer version of that is that cron failed after an OS
update on our RSSAC stats processing box, and I messed up the
re-processing.  Andrew M spotted this, and I've fixed it.

Would you please re-fetch the data for that time period into the GitHub
repo?   NB: The load-time data is not affected.   Any unique sources
data you have is OK, but some of it was missing.

thanks,

Ray